feat: add processing of text statistics with spacy
Add transformer to extract number of sentences, average sentence word
count and number of unique words in the document

Add pipeline to extract metrics from a saved text

Add transformer to read markdown file text

Add chached function to load Spacy model

+++ schema.py
@@ -23,7 +23,7 @@ class ChatMessageResponseBody(BaseModel):
 
 class StatisticsMetrics(BaseModel):
     number_of_sentences: int
-    average_sentence_length: int
+    average_sentence_length: float
     hapax_legomena_count: int
